Clean up portal controller
