--- db/migrate/20210217215722_add_schedule_to_students.rb
@@ -1,5 +1,0 @@
-class AddScheduleToStudents < ActiveRecord::Migration[6.0]
-  def change
-    add_column :students, :schedule, :time
-  end
-end
