--- db/migrate/20210217220216_change_column_name_students.rb
@@ -1,5 +1,0 @@
-class ChangeColumnNameStudents < ActiveRecord::Migration[6.0]
-  def change
-    rename_column :students, :schedule, :date
-  end
-end
