--- db/migrate/20210217220515_add_time_to_students.rb
@@ -1,5 +1,0 @@
-class AddTimeToStudents < ActiveRecord::Migration[6.0]
-  def change
-    rename_column :students, :date, :time_slot
-  end
-end
