--- db/migrate/20210217222226_add_date_to_students.rb
@@ -1,5 +1,0 @@
-class AddDateToStudents < ActiveRecord::Migration[6.0]
-  def change
-    add_column :students, :date_slot, :date
-  end
-end
