--- db/migrate/20210217225653_add_schedule_array_to_students.rb
@@ -1,5 +1,0 @@
-class AddScheduleArrayToStudents < ActiveRecord::Migration[6.0]
-  def change
-    add_column :students, :schedule, :string, array: true, default: []
-  end
-end
