+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_02_17_225653) do
+ActiveRecord::Schema.define(version: 2021_02_16_215117) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -93,9 +93,6 @@ ActiveRecord::Schema.define(version: 2021_02_17_225653) do
     t.string "email"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-    t.time "time_slot"
-    t.date "date_slot"
-    t.string "schedule", default: [], array: true
   end
 
   create_table "users", force: :cascade do |t|
